Makefile: Now less problems with parallel execution and bision's multiple generated files
A rule with multiple targets, which are all created by a single command, does not work as expected when it comes to parallel execution, because it is interpreted as multiple rules, each having one target. Strangely, for implicit rules with multiple targets GNU make assumes that a single invocation of the recipe WILL build all the targets.