Parser: bugfix: Variant type made wrong assumptions about type_info::name()
Description of type_info::name() (from cppreference.com): Returns an implementation defined null-terminated character string containing the name of the type. No guarantees are given, in particular, the returned string can be identical for several types and change between invocations of the same program.

Thus for the same type T, type_info::name() is allowed to return different pointers, a use case that actually occured with clang.

Actually not even the 'solution' introduced by this commit is in accordance with the standard, since for the same type T the string returned by type_info::name() could have different content. But it appears to work for now.

The proper solution would probably be to use std::type_index objects and the std::type_index::operator== function.