No longer also inserting non-local names into global/root scope
That was done only because until now it was not possible to get the globally unique fully qualified name of an Env node. The next few commits will add that thought.

The added test aka specification only specifies what the semantic analyzer already can do. But the IrGen now (with this commit) would have a problem if a nested name of a non-local would have the same name as an outer non-local, because IrGen currently puts all into the outermost scope within LLVM IR.